refactor(logger): reorganize into src/plugins/ structure
